:recycle: refactor: FolderAction 제거
