+++ backend/src/main/java/kernel360/techpick/feature/folder/service/parser/internal/NodeType.java
@@ -1,0 +1,7 @@
+package kernel360.techpick.feature.folder.service.parser.internal;
+
+public enum NodeType {
+	PICK,
+	FOLDER,
+	;
+}
