--- backend/src/main/java/kernel360/techpick/feature/folder/service/parser/structure/NodeType.java
@@ -1,7 +1,0 @@
-package kernel360.techpick.feature.folder.service.parser.structure;
-
-public enum NodeType {
-	PICK,
-	FOLDER,
-	;
-}
